requirements.txt
async-timeout==4.0.3
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.12.1
et-xmlfile==1.1.0
fonttools==4.47.2
google-api-core==2.16.2
google-api-python-client==2.118.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
httplib2==0.22.0
idna==3.6
kiwisolver==1.4.5
matplotlib==3.8.2
numpy==1.26.3
oauthlib==3.2.2
openpyxl==3.1.2
packaging==23.2
pillow==10.2.0
protobuf==4.25.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-ecobee-api==0.0.19
pytz==2024.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
screenlogicpy==0.9.4
six==1.16.0
uritemplate==4.1.1
urllib3==2.2.0
xlrd==2.0.1
XlsxWriter==3.1.9
xlwt==1.3.0
